+++ internal/database/migration/cliutil/downgrade.go
@@ -6,6 +6,10 @@ import (
 
 	"github.com/urfave/cli/v2"
 
+	"github.com/sourcegraph/sourcegraph/internal/database/migration/multiversion"
+	"github.com/sourcegraph/sourcegraph/internal/database/migration/runner"
+	"github.com/sourcegraph/sourcegraph/internal/database/migration/schemas"
+	"github.com/sourcegraph/sourcegraph/internal/database/migration/store"
 	"github.com/sourcegraph/sourcegraph/internal/oobmigration"
 	"github.com/sourcegraph/sourcegraph/internal/oobmigration/migrations"
 	"github.com/sourcegraph/sourcegraph/lib/errors"
@@ -14,10 +18,10 @@ import (
 
 func Downgrade(
 	commandName string,
-	runnerFactory RunnerFactoryWithSchemas,
+	runnerFactory runner.RunnerFactoryWithSchemas,
 	outFactory OutputFactory,
 	registerMigrators func(storeFactory migrations.StoreFactory) oobmigration.RegisterMigratorsFunc,
-	expectedSchemaFactories ...ExpectedSchemaFactory,
+	expectedSchemaFactories ...schemas.ExpectedSchemaFactory,
 ) *cli.Command {
 	fromFlag := &cli.StringFlag{
 		Name:     "from",
@@ -118,7 +122,7 @@ func Downgrade(
 
 		// Find the relevant schema and data migrations to perform (and in what order)
 		// for the given version range.
-		plan, err := planMigration(from, to, versionRange, interrupts)
+		plan, err := multiversion.PlanMigration(from, to, versionRange, interrupts)
 		if err != nil {
 			return err
 		}
@@ -128,9 +132,21 @@ func Downgrade(
 			return err
 		}
 
+		runner, err := runnerFactory(schemas.SchemaNames, schemas.Schemas)
+		if err != nil {
+			return errors.Wrap(err, "new runner")
+		}
+
+		// connect to db and get upgrade readiness state
+		db, err := store.ExtractDatabase(ctx, runner)
+		if err != nil {
+			return errors.Wrap(err, "new db handle")
+		}
+
 		// Perform the downgrade on the configured databases.
-		return runMigration(
+		return multiversion.RunMigration(
 			ctx,
+			db,
 			runnerFactory,
 			plan,
 			privilegedMode,
