+++ internal/database/migration/cliutil/run_oobmigrations.go
@@ -2,22 +2,16 @@ package cliutil
 
 import (
 	"context"
-	"database/sql"
-	"fmt"
-	"os"
 	"sort"
 	"time"
 
 	"github.com/urfave/cli/v2"
 
-	"github.com/sourcegraph/log"
-
 	"github.com/sourcegraph/sourcegraph/internal/database"
-	"github.com/sourcegraph/sourcegraph/internal/database/basestore"
 	"github.com/sourcegraph/sourcegraph/internal/database/migration/schemas"
+	"github.com/sourcegraph/sourcegraph/internal/database/migration/store"
 	"github.com/sourcegraph/sourcegraph/internal/oobmigration"
 	oobmigrations "github.com/sourcegraph/sourcegraph/internal/oobmigration/migrations"
-	"github.com/sourcegraph/sourcegraph/lib/errors"
 	"github.com/sourcegraph/sourcegraph/lib/output"
 )
 
@@ -48,11 +42,11 @@ func RunOutOfBandMigrations(
 		if err != nil {
 			return err
 		}
-		db, err := extractDatabase(ctx, r)
+		db, err := store.ExtractDatabase(ctx, r)
 		if err != nil {
 			return err
 		}
-		registerMigrators := registerMigratorsWithStore(basestoreExtractor{r})
+		registerMigrators := registerMigratorsWithStore(store.BasestoreExtractor{r})
 
 		if err := runOutOfBandMigrations(
 			ctx,
@@ -139,12 +133,12 @@ func runOutOfBandMigrations(
 		}
 	}()
 
-	updateMigrationProgress, cleanup := makeOutOfBandMigrationProgressUpdater(out, ids, animateProgress)
+	updateMigrationProgress, cleanup := oobmigration.MakeProgressUpdater(out, ids, animateProgress)
 	defer cleanup()
 
 	ticker := time.NewTicker(time.Second).C
 	for {
-		migrations, err := getMigrations(ctx, store, ids)
+		migrations, err := store.GetByIDS(ctx, ids)
 		if err != nil {
 			return err
 		}
@@ -174,77 +168,4 @@ func runOutOfBandMigrations(
 		case <-ticker:
 		}
 	}
-}
-
-// makeOutOfBandMigrationProgressUpdater returns a two functions: `update` should be called
-// when the updates to the progress of an out-of-band migration are made and should be reflected
-// in the output; and `cleanup` should be called on defer when the progress object should be
-// disposed.
-func makeOutOfBandMigrationProgressUpdater(out *output.Output, ids []int, animateProgress bool) (
-	update func(i int, m oobmigration.Migration),
-	cleanup func(),
-) {
-	if !animateProgress || shouldDisableProgressAnimation() {
-		update = func(i int, m oobmigration.Migration) {
-			out.WriteLine(output.Linef("", output.StyleReset, "Migration #%d is %.2f%% complete", m.ID, m.Progress*100))
-		}
-		return update, func() {}
-	}
-
-	bars := make([]output.ProgressBar, 0, len(ids))
-	for _, id := range ids {
-		bars = append(bars, output.ProgressBar{
-			Label: fmt.Sprintf("Migration #%d", id),
-			Max:   1.0,
-		})
-	}
-
-	progress := out.Progress(bars, nil)
-	return func(i int, m oobmigration.Migration) { progress.SetValue(i, m.Progress) }, progress.Destroy
-}
-
-// shouldDisableProgressAnimation determines if progress bars should be avoided because the log level
-// will create output that interferes with a stable canvas. In effect, this adds the -disable-animation
-// flag when SRC_LOG_LEVEL is info or debug.
-func shouldDisableProgressAnimation() bool {
-	switch log.Level(os.Getenv(log.EnvLogLevel)) {
-	case log.LevelDebug:
-		return true
-	case log.LevelInfo:
-		return true
-
-	default:
-		return false
-	}
-}
-
-func getMigrations(ctx context.Context, store *oobmigration.Store, ids []int) ([]oobmigration.Migration, error) {
-	migrations := make([]oobmigration.Migration, 0, len(ids))
-	for _, id := range ids {
-		migration, ok, err := store.GetByID(ctx, id)
-		if err != nil {
-			return nil, err
-		}
-		if !ok {
-			return nil, errors.Newf("unknown migration id %d", id)
-		}
-
-		migrations = append(migrations, migration)
-	}
-	sort.Slice(migrations, func(i, j int) bool { return migrations[i].ID < migrations[j].ID })
-
-	return migrations, nil
-}
-
-type basestoreExtractor struct {
-	runner Runner
-}
-
-func (r basestoreExtractor) Store(ctx context.Context, schemaName string) (*basestore.Store, error) {
-	shareableStore, err := extractDB(ctx, r.runner, schemaName)
-	if err != nil {
-		return nil, err
-	}
-
-	return basestore.NewWithHandle(basestore.NewHandleWithDB(log.NoOp(), shareableStore, sql.TxOptions{})), nil
 }
